Added calculate total amount + methods implementing korting berekenen maar blijf nullpointer krijgen.

+++ src/domain/model/kortingStrategies/GeenKorting.java
@@ -8,7 +8,7 @@ public class GeenKorting implements KortingStrategy{
     }
 
     @Override
-    public double calculateTotalReduction(int amount) {
+    public double calculateTotalReduction(double amount) {
         return amount;
     }
 }
